requirements.txt
streamlit
joblib
numpy
opencv-python-headless
scikit-learn
tensorflow
Pillow
